% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-distributions.R
\name{compare_conv_distributions}
\alias{compare_conv_distributions}
\title{Compares paretoconv with poweRlaw distributions}
\usage{
compare_conv_distributions(d1, d2)
}
\arguments{
\item{d1}{Discrete power law distribution of class \code{displ} from the
package \code{poweRlaw}.}

\item{d2}{Equivalent convoluted values for the probability density function
obtained from \code{paretoconv (..., cdf = FALSE)}}
}
\description{
This function extends directly from the \code{poweRlaw} function of the same
name, and compares a discrete power law distribution from the \code{poweRlaw}
package with an equivalent convoluted distribution generated by
\code{paretoconv}.
}
\examples{
\dontrun{
data ('native_american', package = 'poweRlaw')
m1 <- displ$new(native_american$Cas)
m1$setXmin (3)
m1$setPars (estimate_pars(m1))
x <- sort (unique (native_american$Cas))
y <- paretoconv (x = x, a = m1$getPars (), x0 = 3, n = 2,
cdf = FALSE, quiet = FALSE)
compare_conv_distributions (m1, y)
}
}
