% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoconv.R
\name{get_asymp_limit}
\alias{get_asymp_limit}
\title{Get limit of asymptotic convergence}
\usage{
get_asymp_limit(a, n, x0, tol = 0.05, quiet = FALSE)
}
\arguments{
\item{a}{The primary shape parameter of the Pareto distribution (single value
only)}

\item{n}{Number of convolutions (single value only)}

\item{x0}{Lower cut-off point of classical heavy-tailed distribution
(generally obtained emprically with the poweRlaw package).}

\item{tol}{Convergence tolerance (values around 0.05 are typically
sufficient; values < 0.01 will generally not work).}

\item{quiet}{If FALSE, issue progress messages}
}
\value{
Single integer value of limit beyond which PDF may be approximated by
its asymptotic power law
}
\description{
Get limit of asymptotic convergence
}
