% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{pareto_optimise}
\alias{pareto_optimise}
\title{Locally optimal fitting of convoluted Pareto distribution to observed data}
\usage{
pareto_optimise(m, x0 = 1, n = 1, check_non_conv = TRUE,
  quiet = TRUE)
}
\arguments{
\item{m}{A \code{poweRlaw::displ} object containing data to be modelled}

\item{x0}{Initial guess for lower limit of Pareto distribution}

\item{n}{Initial guess for number of convolutions}

\item{check_non_conv}{If TRUE first checks whether a non-convoluted model may
be optimal}

\item{quiet}{If FALSE, progress information is dumped to screen}
}
\value{
Position of the local optimum as quantified by \code{x0} and
\code{n}, along with associated Kolmogorow-Smirnov statistic quantifying
maximal distance from convoluted Pareto Cumulative Distribution Function
(CDF) and empirical CDF of model \code{m}.
}
\description{
Locally optimal fitting of convoluted Pareto distribution to observed data
}
\note{
This function only finds local optima - it is up to the user to ensure
that the given start values are near the global optimum. Values of \code{n = 0}
are NOT searched.
}
