% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-distributions.R
\name{pdf_integral}
\alias{pdf_integral}
\title{Calculates integral of probability density functions for convoluted Pareto
distributions}
\usage{
pdf_integral(a, x0 = 1, n = 0, x0lim = FALSE, quiet = TRUE)
}
\arguments{
\item{a}{Shape parameter of Pareto disitribution}

\item{x0}{Lower cut-off for classic Pareto distribution}

\item{n}{Number of convolutions}

\item{x0lim}{If TRUE, integral is calculated only for values of \code{x>=x0}.}

\item{quiet}{If FALSE, progress is displayed as screen output}
}
\description{
Calculates integral of probability density functions for convoluted Pareto
distributions
}
\note{
This function can take a long time to execute!
}
\examples{
pdf_integral (a=2.1, x0=5, n=0)
}
