% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{pparetoconv}
\alias{pparetoconv}
\title{Estimate probability of observed KS statistic from synthetic models}
\usage{
pparetoconv(m, x0, n, neach = 10, quiet = TRUE)
}
\arguments{
\item{m}{A \code{poweRlaw::displ} object containing data to be modelled}

\item{x0}{Initial guess for lower limit of Pareto distribution}

\item{n}{Initial guess for number of convolutions}

\item{neach}{Number of test statistics to be generated for each different
model}

\item{quiet}{If FALSE, display progress messages on screen}
}
\value{
Single value representing probability of observing given value of
\code{ks}
}
\description{
Estimate probability of observed KS statistic from synthetic models
}
