% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay-int.R
\name{ramsay_int}
\alias{ramsay_int}
\title{Convolution of Pareto distributions for non-integer alpha}
\usage{
ramsay_int(x, a, n)
}
\arguments{
\item{x}{value of independent variable}

\item{a}{The primary shape parameter of the Pareto distribution - alpha in
Ramsay's notation.}

\item{n}{Number of convolutions}
}
\value{
Value for the convolution of two Pareto distributions of shape a at
the value x.
}
\description{
Convolutes multiple Pareto distributions following 'The Distribution of Sums
of Certain I.I.D. Pareto Variates' by Colin Ramsay (Communications in
Statistics - Theory and Methods 35:395-405, 2006).
}
\note{
This version is for integer values of a. Non-Integer values can be
calculated with ramsay_nonint
}
\examples{
ramsay_int (1, 1, 1)
}

