% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsay-nonint.R
\name{ramsay_nonint}
\alias{ramsay_nonint}
\title{Convolution of Pareto distributions for non-integer alpha}
\usage{
ramsay_nonint(x, a, n)
}
\arguments{
\item{x}{value of independent variable}

\item{a}{The primary shape parameter of the Pareto distribution - alpha in
Ramsay's notation.}

\item{n}{Number of convolutions}
}
\value{
Value for the convolution of two Pareto distributions of shape a at
the value x.
}
\description{
Convolution of Pareto distributions for non-integer alpha
}
\note{
This version is for non-integer values of a. Integer values can be
calculated with ramsay_int
}
\examples{
ramsay_nonint (1, 0.9, 1)
}

