% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpl.R
\name{rplconv}
\alias{rplconv}
\title{Generate random deviates for power laws}
\usage{
rplconv(n, x0, alpha, discrete_max = 10000)
}
\arguments{
\item{n}{Number of deviates}

\item{x0}{Traditional cutoff parameter of Pareto distribution}

\item{alpha}{Primary shape parameter of Pareto distribution}

\item{discrete_max}{Upper limit for generation (see \code{poweRlaw::rpldis}
for details)}
}
\value{
Vector of \code{n} random deviates
}
\description{
Generate random deviates for power laws
}
\note{
This is directly cribbed from \code{poweRlaw::rpldis}, modified only to
generate random deviates for all values including 0<x<xmin.
}
