% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoconv-package.R, R/paretoconv.R
\docType{package}
\name{paretoconv}
\alias{paretoconv}
\alias{paretoconv-package}
\title{paretoconv.}
\usage{
paretoconv(x, a, n, cdf = FALSE)
}
\arguments{
\item{x}{value of independent variable (may be a vector)}

\item{a}{The primary shape parameter of the Pareto distribution - alpha in
Ramsay's notation (single value only)}

\item{n}{Number of convolutions (single value only)}

\item{cdf}{If TRUE, returns the cumulative distribution function, other
returns the probability density function.}
}
\value{
Value for the convolution of two Pareto distributions of shape a at
the value x.
}
\description{
paretoconv.

Convolutes multiple Pareto distributions following 
\itemize{
\item For integer shape parameters: 'The Distribution of Sums of Certain
I.I.D.  Pareto Variates' by Colin Ramsay (Communications in Statistics -
Theory and Methods 35:395-405, 2006);
\item For non-integer shape parameters: 'The Distribution of Sums of I.I.D.
Pareto Random Variables with Arbitrary Shape Parameter' by Colin Ramsay
(Communications in Statistics - Theory and Methods 37:2177-2184, 2008).
}
}
\note{
The Pareto distribution may be defined as f(x)=(a/b)(b/x)^(a-1), where
a and b are the primary and secondary shape parameters, respectively. It
presumed here without loss of generality that b=1 and thus f(x)=a x^(1-a).
}
\examples{
paretoconv (1:10, 1, 1)
}
\author{
Mark Padgham
}

