% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-distributions.R
\name{pdf_integral}
\alias{pdf_integral}
\title{Calculates integral of probability density functions for convoluted Pareto
distributions}
\usage{
pdf_integral(a, x0 = 1, n = 0, discrete = TRUE, xmin = 0,
  quiet = TRUE)
}
\arguments{
\item{a}{Shape parameter of Pareto disitribution}

\item{x0}{Lower cut-off for classic Pareto distribution}

\item{n}{Number of convolutions}

\item{discrete}{Calculate integral for discrete (TRUE) or continuous (FALSE)
probability density function}

\item{xmin}{Lower limit of integral (generally 0, 1, or x0)}

\item{quiet}{If FALSE, progress is displayed as screen output}
}
\description{
Calculates integral of probability density functions for convoluted Pareto
distributions
}
\note{
Values for discrete integrals differ from values from continuous
integrals. This function can also take a long time to execute, particularly
for continuous integrals.
}
\examples{
pdf_integral (a=2.1, x0=5, n=0)
}
